Implement field-name path navigation
This introduces the basic path handling which uses the field name to
navigate the YAML structure. This is accomplished by implementing a new
set of `Expression` types that extract subfields from the current
context.

In order to properly test this, a new testing system has been added that
leverages a `go.work` and a new `test` subdirectory which depends on
both this implementation, and the `godog` library. This allows for the
tests to be written more _human-readable_ and holistically done.

Unit tests will be introduced later.