Migrate wildcard expressions to its own type
Previously this was merged into the Field expression, but it's becoming
more apparent that it's logically better to be its own unique type. This
is because the bracket expression and other sub-expressions will require
this logic as well.

This moves the wildcard expression to its own type, and updates code
that previously leveraged the field type. All tests continue to pass.