Implement base filtering logic
The JSONPath specification supports filtering via the `?(...)`
expressions. There's a number of cases that need to be supported with
filtering, so to keep this simple, this starts by just implementing
basic boolean filtering of the values. Features for comparisons,
existence checks, and other types of filtering will be added in
subsequent commits.