Rework YAMLPath grammar
The ANTLR grammar used for the YAMLPath logic was getting messy and
really hard to maintain. Adding in binary expressions for boolean logic
was the final straw, since it lead to precedence issues that would be
easier to resolve by rewriting the grammar since it enables embedding
the precedence in the parse-tree.

This is what's done in this commit. The grammar is rewritten, and the
visitor tree was modified to handle the new parse-tree. Tests continue
to pass.