Add polarity operator support
Currently the YAMLPath implementation does not offer polarity operators
for unary prefix operations. This adds some basic functionality for it.

In the process, the grammar was updated to correct the precedence order
of the math operators, which were unintentionally inverted. The root
expression and invocation expressions were raised in priority as well.