Fix equality with numeric values
Equality was previously being handled as just string-to-string
comparisons (e.g. `"1" == "1"`). This is mostly correct, but it breaks
on numeric values where the representations are different but the
values are identical -- such as scientific notation with decimal
notation (e.g. `1.0 == 1e0`).

This change now parses the values to be treated as true numeric values
and compares the representations instead. This should be more robust
and correct.