Rework YAMLPath collection API
The current YAMLPath API provides back `[]*yaml.Node` objects as the
primary communication mechanism. Although this works, it is actually
rather difficult to work with in order to convert back to standard Go
types. The initial plan was to offer `Eval*` funcs everywhere so that
it can be more convenient to get values back out, but this is ultimately
not the most ideal solution since it bloats the API.

To combat this, the return type has been changed to its own shallow
wrapper now -- the `Collection`, which can also be converted into a
`Decoder` for generic decoding of sequences. This now makes it much
easier for callers to convert to either solitary/singleton values, or
to slices of values via helpers from the Collection type.

In the process, the `Eval` top-level func has been renamed to `Match` to
model the Regex-style API. This is technically a breaking change, but
nobody is even using this library yet since it's still in its infancy
and thus volatile.