Add unit tests for expr package (part 2)
This completes adding full unit tests for the `expr` package.
Previously, testing was largely deferred to the Godog functional
cucumber tests, rather than unit tests -- since it gave a holistic view
of the system. Now that the majority of the functionality is implemented,
it's time to start productionizing this codebase and providing good
unit test coverage.

Some small changes were made in the process so that the arithmetic
operators had named values to choose from, but otherwise the majority
of the changes were adding unit tests, and fixing some behaviors
discovered in those tests.