Refactor exprs to use new expr.Context
The original API took both a `context.Context` along with the nodes for
the current value -- but this ultimately lead to some level of
duplication, as well as full orthogonality between parameters.
Expressions would always either take the root _or_ the current value,
but never both -- meaning it would always be either `ctx` or `nodes`.

A more logical abstraction here is to create a custom context instead,
where the option is always decided from the context itself.