Add support for YAMLPath functions
This large change finally implements function-call support in YAMLPath.
A large part of implementing this requires abstracting out the
`*expr.Context` to avoid a circular dependency. The easiest approach
here is to leverage an `interface` that models the `Context` rather than
relying on the concrete type; but the problem that emerges here is that
`SubContext(...)` returns a `Context`. If it returns an interface, then
it forces `*expr.Context` to return by an interface instead of by
concrete type.

Effectively, this forces either having a weird API that is part-concrete
part-interface, or it's easier to just model everything in terms of this
instead. Ultimately that's how I chose to do this. By changing to an
interface everywhere, it also means all expressions and their related
tests needed to be updated to reflect this change.

Some functions have been added to verify function-call behavior.
Some basic cucumber tests have been added to validate the new functions.
Unit tests will be added in a follow-on commit since the size of this
change has ballooned.