Refactor errors to new package
Most errors within the library were defined under `expr`, but this is now
leading to some import-cycles when trying to leverage the errors in the
function definitions -- since funcs are used from within the expr
package.

To avoid these conflicts, I've decided to promote the errors to their own
shared package that multiple packages can use now. Additionally, they're
now exported in the main `yamlpath` package -- since this makes it
easier to test against.

The errors have been built out to lead to better diagnostics when
testing, so that the exact underlying cause can be better determined
when a test fails.