Add keys() functionality for extracting map keys
JSONPath does not natively support getting the keys from a map, which
is a semi-common operation in queries. There are some extensions with
uncommon syntax using `~`; but since this YAMLPath implementation
supports functions, this chooses to delegate it to `keys()` instead.

A consequence of this design is that it's easier to see all the keys
within a document by using `$..keys()`.