Add custom function support
After some time, the much awaited "custom function" support is finally
here. This enables easy use of reflection to define custom behaviors
that can be invoked within a YAMLPath expression.

This has been implemented as a new `Option` abstraction that can be
provided to the `Compile` function. The arity of the function will be
computed based on the number of arguments and return type of the
reflected function. Function arguments, and return types, can be any
standard Go type, and will be implicitly decoded during evaluation to
make this as flexible as possible.