Merge compliance test coverage into unit test
Currently, unit test coverage is all that is collected and uploaded
during the GitHub Actions workflow -- which means any of the compliance
tests for behaviors are not included in the coverage report. This aims
to change that, by now generating coverage for both the normal test
as well as the `./test/...` Cucumber tests.

This should enable broader coverage of the generated ANTLR parse tree.