Fix YAMLPath list aggregation
The list aggregation definition was erroneously using the literal
rule instead of the aggregation rule. The consequence of this mistake
was that the parser could only accept literal values in list entries,
such as `[1, 2, 3]`, but not more complex aggregations like
`[1, 2, 3, {"a": 1}]`.

This fixes the rule definition to properly accept aggregations, and
updates the generated parser code. By coincidence, no other changes
are necessary for the visitor code, since the visitor always just
handles the raw text of the aggregation as a JSON definition.