Remove filter expressions
The filtering expressions previously borrowed from JSONPath, but are
not the most ergonomic to read or write. Ultimately, this functionality
is fully duplicated by the `where` function, which is more readable and
just as easy to use.

To simplify the grammar, the filter expression is being removed in favor
of the `where` function.