Add initial implementation of union subexpressions
The current implementation of YAMLPath does not enable any union of
multiple fields to be selected. To combat this, this introduces a new
subexpression type that allows for the union of multiple subexpressions
to be selected. This is done by using the pipe operator `|` between
subexpressions.

This will be expanded on in a later change so that it can be done
inline like `$.foo.(@.bar | @.baz)` to select `$.foo.bar` and
`$.foo.baz`. This would be equivalent to `$.foo["bar","baz"]` in
JSONPath.