Refactor root terms
The initial YAMLPath definition exclusively worked with `$` as a root
identifier and `@` as a current context identifier. This also means
that literals are not valid root terms, and cannot be provided as
proper YAMLPath expressions -- only subexpressions.

This is attempting to fix that by extending term invocations to include
literals and invocations. This will, in effect, allow for YAMLPath to
generically support root terms, function invocations without any
specified fields, and literals.

In the future this will also expand so that _any_ arbitrary
subexpression may become a proper expression.