Include source location in reflection output
Reflection currently doesn't indicate the line and column of the
source node. However, this information can be useful in both debugging
and for reporting error when using this tool for analysis.

This adds some small alterations to make it possible to surface this
in the `reflect()` function, and also strips the location when
parsing inline YAML.