Add string functions for regex match/replace
Regex matching is currently handled exclusively in the grammar via the
`~=` operator -- which doesn't read very well, and requires the regex
to be defined within the operational grammar. The positive of this
approach is that regexes can be compiled early for faster execution,
but the downside is that this needs to be defined in the base grammar,
cannot come as the result of an expression from another function or be
data-driven from a YAML configuration node, and is not as
readable as a function call.

As an incremental step forward, this introduces both a `matches` and
`replaceMatches` function that operates on strings. The former has the
same semantics as the `~=` operator, but can come from any arbitrary
expression returning a string -- whether a literal or a separate
YAMLPath. The `replaceMatches` also exists for symmetry with the existing
`replace` function.

Long-term, the `~=` operator will be removed in favor of this new
`matches` func, which also frees up this operator for a new "equivalence"
operator which will repurpose this operator.