Rename `yamlutil` -> `yamlhash`
In the final effort to remove the `*util` "miscellaneous kitchen drawer"
package from this library, the last remaining func in this package is
for hashing. Ultimately, the only real organization for this is to move
it into its own hashing package now -- so the `yamlutil` package is
simply being renamed to be `yamlhash`.