Add more test-cases for compilation
The test-suite of known valid and rejected expressions hasn't been
updated in some time, and this is pulling down coverage information for
the visitor itself. This adds some new values for testing to improve
some coverage.

Long-term, more cucumber tests will be added which will pull this even
higher as well.