Fix coverage computation from cucumber tests
Cucumber tests don't appear to be running on CI; instead we're running
the normal unit tests twice due to an erroneous `./...` in the
`go test` command, when it should be `./test/...`.

This resolves that issue. It also changes the `-coverpkg` to refer to
`./...` since `./...` from the root is the normal library, whereas
`./test/...` is the test `go.mod`.