Update go toolchain version
The GitHub CI runners have specific Go versions cached for use.
Unfortunately the `toolchain` version currently specified is not one
such version, and this is resulting in it downloading the Go version
every time -- and adding an extra 15s to the build.

Using a cached toolchain runner as specified in the `runner-images`
repos should cut down on this cost.