Add navigation functionality
The long-term goal of recent refactorings is to remove the recursive
descent operator $.. and replace it with a `descendants()` function
so that the semantics are easier to read and understand. This is the
first step by introducing the `descendants()` function, along with
a `children()` function, to make it easier to navigate.

In a later commit, the `$..` operator will be removed.